pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-telegram-bot>=21.10",
    "sqlalchemy>=2.0.38",
]

[dependency-groups]
dev = [
    "ruff>=0.9.6",
]
